[DROOLS-6601] WIP -Keep cleaning up/refactoring
